+++ lib/cocoapods/command/install.rb
@@ -30,7 +30,13 @@ module Pod
       def self.options
         [
           ['--repo-update', 'Force running `pod repo update` before install'],
+          ['--deployment', 'Disallow any changes to the Podfile or the Podfile.lock during installation'],
         ].concat(super).reject { |(name, _)| name == '--no-repo-update' }
+      end
+
+      def initialize(argv)
+        super
+        @deployment = argv.flag?('deployment', false)
       end
 
       def run
@@ -38,6 +44,7 @@ module Pod
         installer = installer_for_config
         installer.repo_update = repo_update?(:default => false)
         installer.update = false
+        installer.deployment = @deployment
         installer.install!
       end
     end